Remove unnecessary instance variables
I had originally included them to make things easier in the view, but have removed them and updated the views accordingly. I'm not sure how to get aorund the 2 in lines 14 and 15 in the create action, and need to look into that a little more because it's been a year since I've worked with this at all in depth.