Fix add to collection button and form stuff
I hated the "add to my collection" button. Changing it from "button_to" to "link_to" allowed me to fix the CSS. Everything is less horrible. Also I moved the "unhappy?" text in the edit account form (which I think was created automatically with Devise, but it's been a year since I worked on that part of my site). I might make more changes to the form later.