Fix topbar and picture layout
The topbar I fixed by changing "sign up" to a button. I might switch it to sign in being the button instead. I also styled the button a bit to fit with the color scheme. I also overrode Foundation's grid, which normally puts the last item in a row to the right, which I think works awfully. I had to add the class end to the last ink pic, which required a bit of logicking. Now it's working for both inks and user_inks.